# Refactor querier batch iterator to make it general purpose